Separating unit tests and integration tests
